+++ GameServerLib/Maps/SummonersRift.cs
@@ -19,6 +19,10 @@ namespace LeagueSandbox.GameServer.Maps
 {
     internal class SummonersRift : IMapProperties
     {
+        public bool HasTopLane { get; set; } = true;
+        public bool HasMidLane { get; set; } = true;
+        public bool HasBotLane { get; set; } = true;
+
         private static readonly List<Vector2> BlueTopWaypoints = new List<Vector2>
         {
             new Vector2(917.0f, 1725.0f),
@@ -130,15 +134,6 @@ namespace LeagueSandbox.GameServer.Maps
             MinionSpawnType.MINION_TYPE_CASTER,
             MinionSpawnType.MINION_TYPE_CASTER
         };
-
-        private static readonly Dictionary<TurretType, int[]> TurretItems = new Dictionary<TurretType, int[]>
-        {
-            { TurretType.OUTER_TURRET, new[] { 1500, 1501, 1502, 1503 } },
-            { TurretType.INNER_TURRET, new[] { 1500, 1501, 1502, 1503, 1504 } },
-            { TurretType.INHIBITOR_TURRET, new[] { 1501, 1502, 1503, 1505 } },
-            { TurretType.NEXUS_TURRET, new[] { 1501, 1502, 1503, 1505 } }
-        };
-
         private Game _game;
         private int _cannonMinionCount;
         private int _minionNumber;
@@ -164,45 +159,7 @@ namespace LeagueSandbox.GameServer.Maps
         {
             _game = game;
             _mapData = game.Config.MapData;
-            _nexus = new List<Nexus>();
-            _inhibitors = new Dictionary<TeamId, Dictionary<LaneID, List<Inhibitor>>>
-            {
-                {
-                    TeamId.TEAM_BLUE, new Dictionary<LaneID, List<Inhibitor>>
-                    {
-                        { LaneID.MIDDLE, new List<Inhibitor>() },
-                        { LaneID.TOP, new List<Inhibitor>() },
-                        { LaneID.BOTTOM, new List<Inhibitor>() }
-                    }
-                },
-                {
-                    TeamId.TEAM_PURPLE, new Dictionary<LaneID, List<Inhibitor>>
-                    {
-                        { LaneID.MIDDLE, new List<Inhibitor>() },
-                        { LaneID.TOP, new List<Inhibitor>() },
-                        { LaneID.BOTTOM, new List<Inhibitor>() }
-                    }
-                }
-            };
-            _turrets = new Dictionary<TeamId, Dictionary<LaneID, List<LaneTurret>>>
-            {
-                {
-                    TeamId.TEAM_BLUE, new Dictionary<LaneID, List<LaneTurret>>
-                    {
-                        { LaneID.MIDDLE, new List<LaneTurret>() },
-                        { LaneID.TOP, new List<LaneTurret>() },
-                        { LaneID.BOTTOM, new List<LaneTurret>() }
-                    }
-                },
-                {
-                    TeamId.TEAM_PURPLE, new Dictionary<LaneID, List<LaneTurret>>
-                    {
-                        { LaneID.MIDDLE, new List<LaneTurret>() },
-                        { LaneID.TOP, new List<LaneTurret>() },
-                        { LaneID.BOTTOM, new List<LaneTurret>() }
-                    }
-                }
-            };
+
             _surrenders = new Dictionary<TeamId, SurrenderHandler>
             {
                 { TeamId.TEAM_BLUE, new SurrenderHandler(game, TeamId.TEAM_BLUE, 1200000.0f , 300000.0f , 30.0f) },
@@ -213,6 +170,14 @@ namespace LeagueSandbox.GameServer.Maps
 
         public int[] GetTurretItems(TurretType type)
         {
+            Dictionary<TurretType, int[]> TurretItems = new Dictionary<TurretType, int[]>
+        {
+            { TurretType.OUTER_TURRET, new[] { 1500, 1501, 1502, 1503 } },
+            { TurretType.INNER_TURRET, new[] { 1500, 1501, 1502, 1503, 1504 } },
+            { TurretType.INHIBITOR_TURRET, new[] { 1501, 1502, 1503, 1505 } },
+            { TurretType.NEXUS_TURRET, new[] { 1501, 1502, 1503, 1505 } }
+        };
+
             if (!TurretItems.ContainsKey(type))
             {
                 return null;
@@ -221,91 +186,64 @@ namespace LeagueSandbox.GameServer.Maps
             return TurretItems[type];
         }
 
-        public void AddTurret(MapData.MapObject parentObject, Vector2 position, TeamId team, string teamName, TurretType type, LaneID lane)
+        public string GetTowerModel(TurretType type, TeamId teamId)
         {
             string towerModel = "";
-            if (team == TeamId.TEAM_BLUE)
+            if (teamId == TeamId.TEAM_BLUE)
             {
-                if (type == TurretType.FOUNTAIN_TURRET)
+                switch (type)
                 {
-                    towerModel = "TurretShrine";
-                }
-                else if (type == TurretType.NEXUS_TURRET)
-                {
-                    towerModel = "TurretAngel";
-                }
-                else if (type == TurretType.INHIBITOR_TURRET)
-                {
-                    towerModel = "TurretDragon";
-                }
-                else if (type == TurretType.INNER_TURRET)
-                {
-                    towerModel = "TurretNormal2";
-                }
-                else if (type == TurretType.OUTER_TURRET)
-                {
-                    towerModel = "TurretNormal";
+                    case TurretType.FOUNTAIN_TURRET:
+                        towerModel = "TurretShrine";
+                        break;
+
+                    case TurretType.NEXUS_TURRET:
+                        towerModel = "TurretAngel";
+                        break;
+
+                    case TurretType.INHIBITOR_TURRET:
+                        towerModel = "TurretDragon";
+                        break;
+
+                    case TurretType.INNER_TURRET:
+                        towerModel = "TurretNormal2";
+                        break;
+
+                    case TurretType.OUTER_TURRET:
+                        towerModel = "TurretNormal";
+                        break;
+
                 }
             }
             else
             {
-                if (type == TurretType.FOUNTAIN_TURRET)
+                switch (type)
                 {
-                    towerModel = "TurretShrine";
-                }
-                else if (type == TurretType.NEXUS_TURRET)
-                {
-                    towerModel = "TurretGiant";
-                }
-                // Nexus and Inhib Towers Might be swapped, double check if that's right.
-                else if (type == TurretType.INHIBITOR_TURRET)
-                {
-                    towerModel = "TurretNormal";
-                }
-                else if (type == TurretType.INNER_TURRET)
-                {
-                    towerModel = "TurretWorm2";
-                }
-                else if (type == TurretType.OUTER_TURRET)
-                {
-                    towerModel = "TurretWorm";
+                    case TurretType.FOUNTAIN_TURRET:
+                        towerModel = "TurretShrine";
+                        break;
+
+                    case TurretType.NEXUS_TURRET:
+                        towerModel = "TurretGiant";
+                        break;
+
+                    // Nexus and Inhib Towers Might be swapped, double check if that's right.
+                    case TurretType.INHIBITOR_TURRET:
+                        towerModel = "TurretNormal";
+                        break;
+
+                    case TurretType.INNER_TURRET:
+                        towerModel = "TurretWorm2";
+                        break;
+
+                    case TurretType.OUTER_TURRET:
+                        towerModel = "TurretWorm";
+                        break;
+
                 }
             }
-
-            _turrets[team][lane].Add(new LaneTurret(_game, parentObject.Name + "_A", teamName + towerModel, position, team, type, GetTurretItems(type), 0, lane, parentObject));
+            return towerModel;
         }
-
-        public TurretType GetTurretType(int trueIndex, LaneID lane)
-        {
-            TurretType returnType = TurretType.FOUNTAIN_TURRET;
-
-            if (lane == LaneID.MIDDLE)
-            {
-                if (trueIndex < 3)
-                {
-                    returnType = TurretType.NEXUS_TURRET;
-                    return returnType;
-                }
-
-                trueIndex -= 2;
-            }
-
-            if (trueIndex == 1)
-            {
-                returnType = TurretType.INHIBITOR_TURRET;
-            }
-            else if (trueIndex == 2)
-            {
-                returnType = TurretType.INNER_TURRET;
-            }
-            else if (trueIndex == 3)
-            {
-                returnType = TurretType.OUTER_TURRET;
-            }
-
-            return returnType;
-        }
-
         public void Init()
         {
             // Announcer events
@@ -315,247 +253,13 @@ namespace LeagueSandbox.GameServer.Maps
             _game.Map.AnnouncerEvents.Add(new Announce(_game, _firstSpawnTime, Announces.MINIONS_HAVE_SPAWNED, false)); // Minions have spawned (90 * 1000)
             _game.Map.AnnouncerEvents.Add(new Announce(_game, _firstSpawnTime, Announces.MINIONS_HAVE_SPAWNED2, false)); // Minions have spawned [2] (90 * 1000)
 
-            // TODO: Generate & use exact positions from content files
-
-            //TODO: Unhardcode everything (preferably by reading from content)
-            // These only matter for projectile collisions.
-            var inhibRadius = 214;
-            var nexusRadius = 353;
-            var sightRange = 1700;
-
-            // Below is where we create the buildings.
-
-            // These two are used for fixing any wrongly indexed turrets that are present.
-            List<MapData.MapObject> missedTurrets = new List<MapData.MapObject>();
-
-            foreach (var mapObject in _mapData.MapObjects.Values)
-            {
-                GameObjectTypes objectType = mapObject.GetGameObjectType();
-
-                if (objectType == 0)
-                {
-                    continue;
-                }
-
-                TeamId teamId = mapObject.GetTeamID();
-                LaneID lane = mapObject.GetLaneID();
-                Vector2 position = new Vector2(mapObject.CentralPoint.X, mapObject.CentralPoint.Z);
-                // Models are specific to team.
-                string teamName = mapObject.GetTeamName();
-
-                // Nexus
-                if (objectType == GameObjectTypes.ObjAnimated_HQ)
-                {
-                    _nexus.Add(new Nexus(_game, teamName + "Nexus", teamId, nexusRadius, position, sightRange, Crc32Algorithm.Compute(Encoding.UTF8.GetBytes(mapObject.Name)) | 0xFF000000));
-                }
-                // Inhibitors
-                else if (objectType == GameObjectTypes.ObjAnimated_BarracksDampener)
-                {
-                    _inhibitors[teamId][lane].Add(new Inhibitor(_game, teamName + "Inhibitor", lane, teamId, inhibRadius, position, sightRange, Crc32Algorithm.Compute(Encoding.UTF8.GetBytes(mapObject.Name)) | 0xFF000000));
-                }
-                // Turrets
-                else if (objectType == GameObjectTypes.ObjAIBase_Turret)
-                {
-                    if (mapObject.Name.Contains("Shrine"))
-                    {
-                        AddTurret(mapObject, position, teamId, teamName, TurretType.FOUNTAIN_TURRET, LaneID.MIDDLE);
-                        continue;
-                    }
-
-                    int index = mapObject.ParseIndex();
-
-                    // Failed to find an index in the turret's name, skip it altogether since it would be invalid.
-                    if (index == -1)
-                    {
-                        // TODO: Verify if we should still add them; they would be assigned to lane NONE as a fountain turret.
-                        continue;
-                    }
-
-                    var turretType = GetTurretType(index, lane);
-
-                    if (turretType == TurretType.FOUNTAIN_TURRET)
-                    {
-                        missedTurrets.Add(mapObject);
-                        continue;
-                    }
-
-                    // index - 1 as we need it to start at 0.
-                    AddTurret(mapObject, position, teamId, teamName, turretType, lane);
-                }
-                else if (objectType == GameObjectTypes.ObjBuilding_SpawnPoint)
-                {
-                    _fountains.Add(teamId, new Fountain(_game, teamId, position, 1000));
-                }
-            }
-
-            // Fix missed turrets (this is basically user-error handling).
-            // Could be improved with pattern recognition rather than a set pattern.
-            if (missedTurrets.Count > 0)
-            {
-                foreach (var missed in missedTurrets)
-                {
-                    TeamId teamId = missed.GetTeamID();
-
-                    LaneID lane = missed.GetLaneID();
-                    // the numbers here can unhardcoded when maps become fully scriptable, perhaps with a MapMetadata detailing the maximum turrets for each lane.
-                    int[] maxLaneSizes = new int[] { 5, 3, 3 };
-
-                    Vector2 position = new Vector2(missed.CentralPoint.X, missed.CentralPoint.Z);
-                    // Models are specific to team.
-                    string teamName = missed.GetTeamName();
-                    TurretType turretType = TurretType.FOUNTAIN_TURRET;
-
-                    int index = missed.ParseIndex();
-
-                    // Failed to find an index in the turret's name.
-                    if (index == -1 || lane == LaneID.NONE)
-                    {
-                        continue;
-                    }
-
-                    // Define whether each lane has been filled with the appropriate number of turrets, 
-                    var midFinished = _turrets[teamId][LaneID.MIDDLE].Count - 1 >= maxLaneSizes[0];
-                    var topFinished = _turrets[teamId][LaneID.TOP].Count - 1 >= maxLaneSizes[1];
-                    var botFinished = _turrets[teamId][LaneID.BOTTOM].Count - 1 >= maxLaneSizes[2];
-
-                    // Define the common pattern that the turrets are created in (unsure if this can/should be unhardcoded).
-                    LaneID[] lanePattern = new LaneID[] { LaneID.MIDDLE, LaneID.TOP, LaneID.BOTTOM };
-                    bool[] finishedPattern = new bool[] { midFinished, topFinished, botFinished };
-
-                    // Compensates for the order of the missing turrets by using the distance from the last valid turret index in the same lane.
-                    int missingTurretDistance = index - (_turrets[teamId][lane].Count - 1);
-
-                    // The lanes have values which allow looping through the pattern.
-                    int patternIndex = missingTurretDistance;
-
-                    if (patternIndex >= lanePattern.Length)
-                    {
-                        // remainder from division is the new index.
-                        patternIndex = patternIndex % lanePattern.Length;
-                    }
-                    else if (patternIndex < 0)
-                    {
-                        patternIndex = -patternIndex % lanePattern.Length;
-                    }
-
-                    // Get the lane after the current lane.
-                    LaneID fixLane = lanePattern[patternIndex];
-
-                    // If the next lane has finished, get the next lane from this finished lane.
-                    if (finishedPattern[patternIndex])
-                    {
-                        fixLane = lanePattern[(int)fixLane];
-                    }
-
-                    if (fixLane != LaneID.NONE)
-                    {
-                        var fixLaneTurrets = _turrets[teamId][fixLane];
-                        var firstIndex = fixLaneTurrets[0].ParentObject.ParseIndex();
-                        var lastIndex = 0;
-
-                        if (fixLaneTurrets.Count > 0)
-                        {
-                            lastIndex = fixLaneTurrets[fixLaneTurrets.Count - 1].ParentObject.ParseIndex();
-                        }
-
-                        // Grab the first or last available index.
-                        var fixIndex = 0;
-                        if (firstIndex > 1)
-                        {
-                            fixIndex = 1;
-                        }
-                        else if (lastIndex < maxLaneSizes[patternIndex])
-                        {
-                            fixIndex = lastIndex + 1;
-                        }
-                        else
-                        {
-                            // Skip turrets which cannot fit.
-                            continue;
-                        }
-
-                        turretType = GetTurretType(fixIndex, fixLane);
-
-                        // If it still failed, it is invalid.
-                        if (turretType != TurretType.FOUNTAIN_TURRET)
-                        {
-                            AddTurret(missed, position, teamId, teamName, turretType, fixLane);
-                        }
-                    }
-                }
-            }
-            var teamInhibitors = new Dictionary<TeamId, List<Inhibitor>>
-            {
-                { TeamId.TEAM_BLUE, new List<Inhibitor>() },
-                { TeamId.TEAM_PURPLE, new List<Inhibitor>() }
-            };
-
-            var teams = teamInhibitors.Keys.ToList();
-            foreach (var team in teams)
-            {
-                _inhibitors[team].Values.ToList().ForEach(l => teamInhibitors[team].AddRange(l));
-            }
-
-            foreach (var nexus in _nexus)
-            {
-                // Adds Protection to Nexus
-                _game.ProtectionManager.AddProtection
-                (
-                    nexus,
-                    _turrets[nexus.Team][LaneID.MIDDLE].FindAll(turret => turret.Type == TurretType.NEXUS_TURRET).ToArray(),
-                    teamInhibitors[nexus.Team].ToArray()
-                );
-
-                // Adds Nexus
-                _game.ObjectManager.AddObject(nexus);
-            }
-
-            // Iterate through all inhibitors for both teams.
-            List<Inhibitor> allInhibitors = new List<Inhibitor>();
-            allInhibitors.AddRange(teamInhibitors[TeamId.TEAM_BLUE]);
-            allInhibitors.AddRange(teamInhibitors[TeamId.TEAM_PURPLE]);
-
-            foreach (var inhibitor in allInhibitors)
-            {
-                var inhibitorTurret = _turrets[inhibitor.Team][inhibitor.Lane].First(turret => turret.Type == TurretType.INHIBITOR_TURRET);
-
-                // Adds Protection to Inhibitors
-                if (inhibitorTurret != null)
-                {
-                    // Depends on the first available inhibitor turret.
-                    _game.ProtectionManager.AddProtection(inhibitor, false, inhibitorTurret);
-                }
-
-                // Adds Inhibitors
-                _game.ObjectManager.AddObject(inhibitor);
-
-                // Adds Protection to Turrets
-                foreach (var turret in _turrets[inhibitor.Team][inhibitor.Lane])
-                {
-                    if (turret.Type == TurretType.NEXUS_TURRET)
-                    {
-                        _game.ProtectionManager.AddProtection(turret, false, _inhibitors[inhibitor.Team][inhibitor.Lane].ToArray());
-                    }
-                    else if (turret.Type == TurretType.INHIBITOR_TURRET)
-                    {
-                        _game.ProtectionManager.AddProtection(turret, false, _turrets[inhibitor.Team][inhibitor.Lane].First(dependTurret => dependTurret.Type == TurretType.INNER_TURRET));
-                    }
-                    else if (turret.Type == TurretType.INNER_TURRET)
-                    {
-                        _game.ProtectionManager.AddProtection(turret, false, _turrets[inhibitor.Team][inhibitor.Lane].First(dependTurret => dependTurret.Type == TurretType.OUTER_TURRET));
-                    }
-
-                    // Adds Turrets
-                    _game.ObjectManager.AddObject(turret);
-                }
-            }
-
             //Map props
             _game.ObjectManager.AddObject(new LevelProp(_game, new Vector2(12465.0f, 14422.257f), 101.0f, new Vector3(0.0f, 0.0f, 0.0f), 0.0f, 0.0f, "LevelProp_Yonkey", "Yonkey"));
             _game.ObjectManager.AddObject(new LevelProp(_game, new Vector2(-76.0f, 1769.1589f), 94.0f, new Vector3(0.0f, 0.0f, 0.0f), 0.0f, 0.0f, "LevelProp_Yonkey1", "Yonkey"));
             _game.ObjectManager.AddObject(new LevelProp(_game, new Vector2(13374.17f, 14245.673f), 194.9741f, new Vector3(224.0f, 33.33f, 0.0f), 0.0f, -44.44f, "LevelProp_ShopMale", "ShopMale"));
             _game.ObjectManager.AddObject(new LevelProp(_game, new Vector2(-99.5613f, 855.6632f), 191.4039f, new Vector3(158.0f, 0.0f, 0.0f), 0.0f, 0.0f, "LevelProp_ShopMale1", "ShopMale"));
         }
+
         public void Update(float diff)
         {
             if (_game.GameTime >= 120 * 1000)
@@ -820,7 +524,7 @@ namespace LeagueSandbox.GameServer.Maps
                         waypoint = BlueBotWaypoints;
                     }
                 }
-                else if (opposed_team == TeamId.TEAM_BLUE) 
+                else if (opposed_team == TeamId.TEAM_BLUE)
                 {
                     if (lane == LaneID.TOP)
                     {
